+++ servo/utilities/__init__.py
@@ -1,6 +1,5 @@
 from .strings import *
 from .duration_str import *
 from .pydantic import *
-from .signal_handler import *
 from .yaml import *
 from .key_paths import *