+++ CHANGELOG.md
@@ -25,6 +25,7 @@ versioned in accordance with [Semantic Versioning](https://semver.org/spec/v2.0.
 - The check command now supports a verbose and non-verbose modes for outputting multiple checks from a connector.
 - The version command can now output version data for connectors.
 - The connectors command now outputs names for displaying aliased configurations.
+- The servo now runs asynchronously on top of asyncio.
 
 ### Changed
 - The reporting interval is now configurable `VegetaConfiguration`.